requirements.txt

ipykernel
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
langchain==0.2.5
langchain-community==0.2.5
langchain-groq==0.1.6
langchain-text-splitters==0.2.1
pypdf==4.2.0
sentence-transformers==3.0.1
chromadb==0.5.3
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.5.0